# tune: optimize hot code paths