More possible optimizations using memoization, on top of improvements in #3349

function statistics before optimizations:

| :fn                                      | :sum      | :calls |
| ---------------------------------------- | --------- | ------ |
| rems.permissions/whitelist               | 1275,58ms | 72128  |
| rems.permissions/blacklist               | 153,97ms  | 9081   |
| rems.permissions/update-role-permissions | 42,15ms   | 9046   |
| rems.db.applications/update-user-roles   | 469,02ms  | 2      |

and after optimization:

| :fn                                      | :sum     | :calls |
| ---------------------------------------- | -------- | ------ |
| rems.permissions/whitelist               | 375,57ms | 72128  |
| rems.permissions/blacklist               | 58,47ms  | 9081   |
| rems.permissions/update-role-permissions | 32,50ms  | 9046   |
| rems.db.applications/update-user-roles   | 179,14ms | 2      |

# Checklist for author

Remove items that aren't applicable, check items that are done.

## Reviewability
- [ ] Link to issue

## Follow-up
- [ ] New tasks are created for pending or remaining tasks